# lib/api-zod/src/generated/types/appointmentCreated.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DEV-001 Thyroid Pre-Visit Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from './appointment';
import type { AppointmentCreatedLink } from './appointmentCreatedLink';

export interface AppointmentCreated {
  appointment: Appointment;
  link?: AppointmentCreatedLink;
  possibleDuplicate?: boolean;
}